refactor: merge into devMiddleware
