--- packages/service/src/server/middlewares/dev/shuvi-dev-middleware/utils/ready.ts
@@ -1,8 +1,0 @@
-import { IContext, ICallback } from '../types';
-
-export default function ready(context: IContext, callback: ICallback) {
-  if (context.state) {
-    return callback(context.stats);
-  }
-  context.callbacks.push(callback);
-}
